+++ server/src/tangible/__init__.py
@@ -1,3 +1,3 @@
 """Tangible server package."""
 
-__version__ = "0.17.19"
+__version__ = "0.17.20"